Transaction analysis task is updated

+++ DLTE-Java/DLTE-Java-003/src/main/java/store/oops/transactionClass.java
@@ -11,7 +11,7 @@ import java.util.Date;
 @NoArgsConstructor
 public class transactionClass {
     private Date dateOfTransaction;
-    private int transactionAmount;
+    private Double transactionAmount;
     private String reciept;
     private String remarks;
